Adds Login to App Container
