+++ src/component/Routes.js
@@ -5,13 +5,12 @@ import AddsCourse from '../container/AddsCourse';
 import AuthorCourses from '../container/AuthorCourses';
 import LogoutContainer from '../container/LogoutContainer';
 import AppContainer from '../container/AppContainer';
-import LoginContainer from '../container/LoginContainer';
 
 const Routes = () => (
   <BrowserRouter>
     <Switch>
       <Route exact path="/" component={AppContainer} />
-      <Route exact path="/signin" component={LoginContainer} />
+      <Route exact path="/signin" component={AppContainer} />
       <Route exact path="/signup" component={SignupContainer} />
       <Route exact path="/addCourse" component={AddsCourse} />
       <Route exact path="/authorCourses" component={AuthorCourses} />
