+++ src/container/SignupContainer.js
@@ -43,18 +43,22 @@ const SignupContainer = () => {
 
   if (user.user.length === 0) {
     return (
-      <div className="signupContainer">
-        <input type="text" name="firstname" placeholder="First Name" onChange={e => handleChange(e)} />
-        <input type="text" name="lastname" placeholder="Last Name" onChange={e => handleChange(e)} />
-        <input type="text" name="username" placeholder="Username" onChange={e => handleChange(e)} />
-        <input type="email" name="email" placeholder="Email" onChange={e => handleChange(e)} />
-        <select name="usertype" defaultValue="user" onChange={e => handleChange(e)}>
-          <option value="user">User</option>
-          <option value="author">Author</option>
-        </select>
-        <input type="password" name="password" placeholder="password" onChange={e => handleChange(e)} />
-        <input type="submit" value="signup" onClick={() => handleSubmit()} />
-        <Link to="/signin/">SIGNIN</Link>
+      <div className="background signupContainer">
+        <div className="signupForm">
+          <input className="inputField" type="text" name="firstname" placeholder="First Name" onChange={e => handleChange(e)} />
+          <input className="inputField" type="text" name="lastname" placeholder="Last Name" onChange={e => handleChange(e)} />
+          <input className="inputField" type="text" name="username" placeholder="Username" onChange={e => handleChange(e)} />
+          <input className="inputField" type="email" name="email" placeholder="Email" onChange={e => handleChange(e)} />
+          <select className="inputField" name="usertype" defaultValue="user" onChange={e => handleChange(e)}>
+            <option value="user">User</option>
+            <option value="author">Author</option>
+          </select>
+          <input className="inputField" type="password" name="password" placeholder="password" onChange={e => handleChange(e)} />
+          <input className="submitField" type="submit" value="signup" onClick={() => handleSubmit()} />
+        </div>
+        <div>
+          <Link className="signinLink" to="/signin/">SIGNIN</Link>
+        </div>
       </div>
     );
   }
