+++ src/styles/index.css
@@ -1,5 +1,4 @@
 @import "reset.css";
-@import "loginContainer.css";
 @import "appContainer.css";
 @import "signupContainer.css";
 @import "courseContainer.css";
