+++ src/styles/signupContainer.css
@@ -1,16 +1,30 @@
 .signupContainer {
-  display: grid;
-  background-color: lightblue;
-  justify-items: center;
-  align-content: center;
-  width: 100%;
-  height: 100%;
-  grid-row-gap: 1.5rem;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
 }
 
-.signupContainer > select,
-.signupContainer > input {
-  height: 30px;
-  width: 80%;
-  border-radius: 5px;
+.signupForm {
+  margin-top: 15px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.signinLink {
+  margin-bottom: 15px;
+  background-color: #fa5943;
+  width: 220px;
+  height: 33px;
+  border-radius: 50px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: white;
+  font-weight: bold;
+  font-size: 16px;
+  text-decoration: none;
 }
